backend/requirements.txt
fastapi
uvicorn
sqlalchemy
sentence-transformers
faiss-cpu
numpy
python-dotenv
yfinance
langchain_pinecone
openai
python-dotenv
langchain-community